+++ apps/commerce-streamer/src/main/java/com/loopers/batch/job/MonthlyProductRankJob.java
@@ -1,6 +1,5 @@
-package com.loopers.batch.config;
+package com.loopers.batch.job;
 
-import java.time.DayOfWeek;
 import java.time.LocalDate;
 
 import org.springframework.batch.core.Job;
@@ -15,7 +14,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 
 import com.loopers.batch.application.ProductRankProcessor;
-import com.loopers.batch.domain.MvProductRankWeekly;
+import com.loopers.batch.domain.MvProductRankMonthly;
 import com.loopers.batch.infrastructure.ProductRankDailyReader;
 import com.loopers.batch.infrastructure.ProductRankWriter;
 import com.loopers.domain.ranking.MvProductRankDaily;
@@ -25,7 +24,7 @@ import lombok.RequiredArgsConstructor;
 @Configuration
 @RequiredArgsConstructor
 @EnableBatchProcessing
-public class WeeklyProductRankBatchConfig {
+public class MonthlyProductRankJob {
     private final JobRepository jobRepository;
     private final PlatformTransactionManager transactionManager;
 
@@ -34,24 +33,24 @@ public class WeeklyProductRankBatchConfig {
     private final ProductRankWriter writer;
 
     @Bean
-    public Job weeklyProductRankJob() {
-        return new JobBuilder("weeklyRankJob", jobRepository)
-                .start(weeklyRankStep())
+    public Job monthlyProductRankJob() {
+        return new JobBuilder("monthlyRankJob", jobRepository)
+                .start(monthlyRankStep())
                 .build();
     }
 
     @Bean
-    public Step weeklyRankStep() {
-        LocalDate startDate = LocalDate.now().minusWeeks(1).with(DayOfWeek.MONDAY);
-        LocalDate endDate = startDate.with(DayOfWeek.SUNDAY);
+    public Step monthlyRankStep() {
+        LocalDate startDate = LocalDate.now().minusMonths(1).withDayOfMonth(1);
+        LocalDate endDate = startDate.withDayOfMonth(startDate.lengthOfMonth());
 
         JpaPagingItemReader<MvProductRankDaily> itemReader = reader.getPagingItemReader(startDate, endDate);
 
-        return new StepBuilder("weeklyRankStep", jobRepository)
-                .<MvProductRankDaily, MvProductRankWeekly>chunk(1000, transactionManager)
+        return new StepBuilder("monthlyRankStep", jobRepository)
+                .<MvProductRankDaily, MvProductRankMonthly>chunk(1000, transactionManager)
                 .reader(itemReader)
-                .processor(processor.processWeekly(startDate, endDate))
-                .writer(writer.writeWeekly())
+                .processor(processor.processMonthly(startDate, endDate))
+                .writer(writer.writeMonthly())
                 .build();
     }
 }
