+++ apps/commerce-streamer/src/main/java/com/loopers/batch/job/WeeklyProductRankJob.java
@@ -1,5 +1,6 @@
-package com.loopers.batch.config;
+package com.loopers.batch.job;
 
+import java.time.DayOfWeek;
 import java.time.LocalDate;
 
 import org.springframework.batch.core.Job;
@@ -14,7 +15,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
 
 import com.loopers.batch.application.ProductRankProcessor;
-import com.loopers.batch.domain.MvProductRankMonthly;
+import com.loopers.batch.domain.MvProductRankWeekly;
 import com.loopers.batch.infrastructure.ProductRankDailyReader;
 import com.loopers.batch.infrastructure.ProductRankWriter;
 import com.loopers.domain.ranking.MvProductRankDaily;
@@ -24,7 +25,7 @@ import lombok.RequiredArgsConstructor;
 @Configuration
 @RequiredArgsConstructor
 @EnableBatchProcessing
-public class MonthlyProductRankBatchConfig {
+public class WeeklyProductRankJob {
     private final JobRepository jobRepository;
     private final PlatformTransactionManager transactionManager;
 
@@ -33,24 +34,24 @@ public class MonthlyProductRankBatchConfig {
     private final ProductRankWriter writer;
 
     @Bean
-    public Job monthlyProductRankJob() {
-        return new JobBuilder("monthlyRankJob", jobRepository)
-                .start(monthlyRankStep())
+    public Job weeklyProductRankJob() {
+        return new JobBuilder("weeklyRankJob", jobRepository)
+                .start(weeklyRankStep())
                 .build();
     }
 
     @Bean
-    public Step monthlyRankStep() {
-        LocalDate startDate = LocalDate.now().minusMonths(1).withDayOfMonth(1);
-        LocalDate endDate = startDate.withDayOfMonth(startDate.lengthOfMonth());
+    public Step weeklyRankStep() {
+        LocalDate startDate = LocalDate.now().minusWeeks(1).with(DayOfWeek.MONDAY);
+        LocalDate endDate = startDate.with(DayOfWeek.SUNDAY);
 
         JpaPagingItemReader<MvProductRankDaily> itemReader = reader.getPagingItemReader(startDate, endDate);
 
-        return new StepBuilder("monthlyRankStep", jobRepository)
-                .<MvProductRankDaily, MvProductRankMonthly>chunk(1000, transactionManager)
+        return new StepBuilder("weeklyRankStep", jobRepository)
+                .<MvProductRankDaily, MvProductRankWeekly>chunk(1000, transactionManager)
                 .reader(itemReader)
-                .processor(processor.processMonthly(startDate, endDate))
-                .writer(writer.writeMonthly())
+                .processor(processor.processWeekly(startDate, endDate))
+                .writer(writer.writeWeekly())
                 .build();
     }
 }
