+++ apps/commerce-streamer/src/main/java/com/loopers/batch/scheduler/ProductRankBatchScheduler.java
@@ -1,4 +1,4 @@
-package com.loopers.batch.config;
+package com.loopers.batch.scheduler;
 
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParameters;
